CMS/requirements.txt
aiohttp==3.9.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorclass==2.2.2
colorlog==6.8.2
croniter==2.0.3
decorator==5.1.1
Django==5.0.1
django-colorfield==0.11.0
django-extensions==3.2.3
django-modeladmin-reorder==0.3.1
django-picklefield==3.2
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
executing==2.0.1
flower==2.0.1
frozenlist==1.4.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
ipython==8.21.0
jedi==0.19.1
kafka-python==1.4.7
kombu==5.3.7
matplotlib-inline==0.1.6
mode==4.3.2
multidict==6.0.5
mypy-extensions==1.0.0
opentracing==1.3.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg==3.1.17
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
robinhood-aiokafka==1.1.6
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
terminaltables==3.1.10
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
ulid-py==1.1.0
uritemplate==4.1.1
urllib3==2.2.0
venusian==1.2.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
